feat: add sentry
