+++ .env.example
@@ -44,8 +44,18 @@ DB_SCHEMA=public
 
 # Prisma database connection
 
-DATABASE_URL=postgresql://${POSTGRES_USER}:${POSTGRES_PASSWORD}@${DB_HOST}:${DB_PORT}/${POSTGRES_DB}?schema=${DB_SCHEMA}&sslmode=prefer&connection_limit=400&pool_timeout=30
+DATABASE_URL=postgresql://${POSTGRES_USER}:${POSTGRES_PASSWORD}@${DB_HOST}:${DB_PORT}/${POSTGRES_DB}?schema=${DB_SCHEMA} &
+sslmode=prefer &
+connection_limit=400 &
+pool_timeout=30
 
 # Redis database connection
 
 REDIS_URL=redis://localhost:6379
+
+# Sentry
+# Sentry
+SENTRY_AUTH_TOKEN=
+SENTRY_ORG=
+SENTRY_PROJECT=
+NEXT_PUBLIC_SENTRY_DSN=
