+++ Dockerfile
@@ -29,6 +29,14 @@ ARG NEXT_PUBLIC_OUR_DOMAIN
 ENV NEXT_PUBLIC_OUR_DOMAIN ${NEXT_PUBLIC_OUR_DOMAIN}
 ARG NEXT_PUBLIC_TWITTER_LINK
 ENV NEXT_PUBLIC_TWITTER_LINK ${NEXT_PUBLIC_TWITTER_LINK}
+ARG SENTRY_ORG
+ENV SENTRY_ORG ${SENTRY_ORG}
+ARG SENTRY_PROJECT
+ENV SENTRY_PROJECT ${SENTRY_PROJECT}
+ARG NEXT_PUBLIC_SENTRY_DSN
+ENV NEXT_PUBLIC_SENTRY_DSN ${NEXT_PUBLIC_SENTRY_DSN}
+ARG SENTRY_AUTH_TOKEN
+ENV SENTRY_AUTH_TOKEN ${SENTRY_AUTH_TOKEN}
 
 ENV NEXT_TELEMETRY_DISABLED 1
 ENV NODE_ENV production
