+++ package.json
@@ -145,6 +145,8 @@
     "@iconify/tailwind": "^0.1.2",
     "@next/bundle-analyzer": "13.3.1",
     "@playwright/test": "1.33.0",
+    "@sentry/nextjs": "7.50.0",
+    "@sentry/webpack-plugin": "1.20.1",
     "@trivago/prettier-plugin-sort-imports": "4.1.1",
     "@types/canvas-confetti": "^1.6.0",
     "@types/chroma-js": "^2.4.0",
