+++ sentry.client.config.ts
@@ -1,0 +1,12 @@
+import * as Sentry from "@sentry/nextjs"
+
+const SENTRY_DSN = process.env.NEXT_PUBLIC_SENTRY_DSN
+
+Sentry.init({
+  dsn: SENTRY_DSN,
+  tracesSampleRate: 0.1,
+  replaysSessionSampleRate: 0.1,
+  replaysOnErrorSampleRate: 0.1,
+  // eslint-disable-next-line import/namespace
+  integrations: [new Sentry.Replay()],
+})
