+++ sentry.properties
@@ -1,0 +1,4 @@
+defaults.url=https://sentry.io/
+defaults.org=
+defaults.project=
+cli.executable=../../node_modules/@sentry/cli/bin/sentry-cli
