+++ sentry.server.config.ts
@@ -1,0 +1,8 @@
+import * as Sentry from "@sentry/nextjs"
+
+const SENTRY_DSN = process.env.NEXT_PUBLIC_SENTRY_DSN
+
+Sentry.init({
+  dsn: SENTRY_DSN,
+  tracesSampleRate: 0.1,
+})
